LFU-813-1 gpio: adp5585: Reset alternate function pins to GPIO
When warm reboot happens, the ADP5585 is not reset. So its registers keep
value before reset. Since kernel is using PWM function for GPIO4 (R3), it
causes problem in u-boot to use this pin as GPIO.
So reset pins C4/R4/R3/R0 with alternate function to GPIO to align with
HW default value.

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>